refactor Pause(exceptions): migrate exception classes to domain layer
- Move PauserException and all child exceptions to domain/exception package
- Add comprehensive Javadoc to all exception classes
- Update imports in Pauser, TlsPauser, TargetSelector, and test classes
- Keep original class names (no renaming to avoid unnecessary complexity)

Note: Initially planned to rename PauserException to DomainException,
but decided to keep the original name as throwing base exception classes
directly is considered an anti-pattern. Base exceptions should only be
used for catching, not throwing.
