--- lib/src/main/java/com/scalar/admin/kubernetes/PauseFailedException.java
@@ -1,11 +1,0 @@
-package com.scalar.admin.kubernetes;
-
-public class PauseFailedException extends PauserException {
-  public PauseFailedException(String message) {
-    super(message);
-  }
-
-  public PauseFailedException(String message, Throwable cause) {
-    super(message, cause);
-  }
-}
