--- lib/src/main/java/com/scalar/admin/kubernetes/PauserException.java
@@ -1,11 +1,0 @@
-package com.scalar.admin.kubernetes;
-
-public class PauserException extends Exception {
-  public PauserException(String message) {
-    super(message);
-  }
-
-  public PauserException(String message, Throwable cause) {
-    super(message, cause);
-  }
-}
