--- lib/src/main/java/com/scalar/admin/kubernetes/StatusCheckFailedException.java
@@ -1,11 +1,0 @@
-package com.scalar.admin.kubernetes;
-
-public class StatusCheckFailedException extends PauserException {
-  public StatusCheckFailedException(String message) {
-    super(message);
-  }
-
-  public StatusCheckFailedException(String message, Throwable cause) {
-    super(message, cause);
-  }
-}
