--- lib/src/main/java/com/scalar/admin/kubernetes/StatusUnmatchedException.java
@@ -1,11 +1,0 @@
-package com.scalar.admin.kubernetes;
-
-public class StatusUnmatchedException extends PauserException {
-  public StatusUnmatchedException(String message) {
-    super(message);
-  }
-
-  public StatusUnmatchedException(String message, Throwable cause) {
-    super(message, cause);
-  }
-}
