--- lib/src/main/java/com/scalar/admin/kubernetes/UnpauseFailedException.java
@@ -1,11 +1,0 @@
-package com.scalar.admin.kubernetes;
-
-public class UnpauseFailedException extends PauserException {
-  public UnpauseFailedException(String message) {
-    super(message);
-  }
-
-  public UnpauseFailedException(String message, Throwable cause) {
-    super(message, cause);
-  }
-}
